added an option to output Google Maps javascript to make world map out of the traceroute output. Example output available at the following location: http://www.foo.be/traceroute-circl/test.html (as need a Google Maps API key to display the result).
Concerning the geolocation lookup, this is limited to the announced
country for a specific whois record entry. So you might find
discrepancies between the real location of the IP address and from
where it is announced. But this is already giving a good overview
between ISP/world scale.

The country lookup is done via a public API of Google without
requiring API key.

TODO :

- Pre-generated list of countries Lat/Long marker
- Improving accuracy of the geolocation lookup