Experimental BGP Ranking output added
BGP Ranking is an experimental services from CIRCL to get a rank
about the number of matches against public blacklist for an ASN.
The minimal value is 1 (no match) to a floating value increasing
with the number of matches. The services is also displaying
the number of match compared to the total public blacklist used.

For more info: http://www.terena.org/activities/tf-csirt/meeting32/dulaunoy-bgpranking.pdf