Add support for IPv6 PTRs
This adds support for resolving IPv6 addresses by using Net::DNS (which,
for reverse lookups, guesses the address type automatically; for the
forward lookup, we guess the query type in a very hackish way by
checking whether a colon appeared in the original address...).

Net::DNS is not a new dependency; it's already used by
Net::Abuse::Utils.